handlers/paste: Remove max character limit
Paste services like 0x0 are mildly frustrating with how they handle
files, it includes file extensions, resulting in FireFox attempting to
download a plain text file that would be better off being displayed in
the browser itself.

Removing this opens up the possibility for tools and communities to
default to using pastebins using this service instead of 0x0.